Updated PCB design
The PCB design has been updated and is now relatively complete. The hardware is fully MIT compliant, as all libraries used are under MIT license, and the design is entirely my own. Now it's time to round the tracks. The schematic has been lightly updated too based on the PCB design (changing a few connections to make routing the board simpler).