Right hand side mostly completed
The right hand side of the keyboard is pretty much complete, and after a final checkover of everything I think it's ready for production. I'm going to clean up the left hand side board design and schematic to match the right hand schematic style, since I based the left hand side on v0.1, and the right hand side was done from scratch.